[core] *TileMonitor => *TileSource
