+++ src/mbgl/annotation/annotation_tile.hpp
@@ -1,10 +1,6 @@
 #pragma once
 
 #include <mbgl/tile/geometry_tile.hpp>
-#include <mbgl/tile/tile_id.hpp>
-
-#include <map>
-#include <unordered_map>
 
 namespace mbgl {
 
@@ -40,24 +36,7 @@ class AnnotationTile : public GeometryTile {
 public:
     util::ptr<GeometryTileLayer> getLayer(const std::string&) const override;
 
-    std::map<std::string, util::ptr<AnnotationTileLayer>> layers;
-};
-
-class AnnotationManager;
-
-class AnnotationTileMonitor : public GeometryTileMonitor {
-public:
-    AnnotationTileMonitor(const OverscaledTileID&, AnnotationManager&);
-    ~AnnotationTileMonitor();
-
-    void update(std::unique_ptr<GeometryTile>);
-    std::unique_ptr<AsyncRequest> monitorTile(const GeometryTileMonitor::Callback&) override;
-
-    OverscaledTileID tileID;
-
-private:
-    AnnotationManager& annotationManager;
-    GeometryTileMonitor::Callback callback;
+    std::unordered_map<std::string, util::ptr<AnnotationTileLayer>> layers;
 };
 
 } // namespace mbgl
